modified for consistency.
made changes regarding car expenses.
header files were moved into header file. names were changed to maintain consistency. printf() changed to remove warning.

+++ header/car_expenses.h
@@ -1,4 +1,4 @@
-#include "../trip.h"
+#include "trip.h"
 
 // bundles total costs with reimbursed costs
 typedef struct allCosts
@@ -7,7 +7,7 @@ typedef struct allCosts
     int allowed;
 } allCosts;
 
-float calculateCarExpenses(struct Trip *tripPointer);
+int calculateCarExpenses(struct Trip *tripPointer);
 float getValidFloat(char msg[]);
 
 void privateCar(struct allCosts *costs);
